# swagger.json
{
    "openapi": "3.0.4",
    "info": {
      "title": "API de Filmes",
      "description": "API para gerenciamento de filmes",
      "version": "1.0.0"
    },
    "paths": {
      "/movies": {
        "get": {
          "tags": ["Movies"],
          "summary": "Obter todos os filmes",
          "responses": {
            "200": {
              "description": "Retorna uma lista de filmes",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Movie"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/movies/{id}": {
        "put": {
          "tags": ["Movies"],
          "summary": "Atualizar um filme existente",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "description": "ID do filme a ser atualizado",
              "schema": {
                "type": "integer"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MovieInput"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Filme atualizado com sucesso"
            },
            "404": {
              "description": "Filme não encontrado"
            },
            "500": {
              "description": "Erro interno do servidor"
            }
          }
        },
        "delete": {
          "tags": ["Movies"],
          "summary": "Excluir um filme",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "description": "ID do filme a ser excluido",
              "schema": {
                "type": "integer"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Filme removido com sucesso"
            },
            "404": {
              "description": "Filme não foi encontrado"
            },
            "500": {
              "description": "Não foi possivel remover o filme"
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
        "Movie": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "description": "ID do filme"
            },
            "title": {
              "type": "string",
              "description": "Título do filme"
            },
            "genre_id": {
              "type": "integer",
              "description": "Gênero do filme"
            },
            "genre": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer",
                  "description": "ID do gênero"
                },
                "name": {
                  "type": "string",
                  "description": "Nome do gênero"
                }
              }
            },
            "language_id": {
              "type": "integer",
              "description": "Idioma do filme"
            },
            "language": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer",
                  "description": "ID do idioma"
                },
                "name": {
                  "type": "string",
                  "description": "Nome do idioma"
                }
              }
            },
            "oscar_count": {
              "type": "integer",
              "description": "Quantidade de Oscars ganhos pelo filme"
            },
            "release_date": {
              "type": "string",
              "format": "date",
              "description": "Data de lançamento do filme"
            }
          }
        },
        "MovieInput": {
          "type": "object",
          "required": [
            "title",
            "genre_id",
            "language_id",
            "oscar_count",
            "release_date"
          ],
          "properties": {
            "title": {
              "type": "string",
              "description": "Título do filme"
            },
            "genre_id": {
              "type": "integer",
              "description": "ID do gênero do filme"
            },
            "language_id": {
              "type": "integer",
              "description": "ID do idioma do filme"
            },
            "oscar_count": {
              "type": "integer",
              "description": "Quantidade de Oscars ganhos pelo filme"
            },
            "release_date": {
              "type": "string",
              "format": "date",
              "description": "Data de lançamento do filme"
            }
          }
        }
      }
    }
  }